+++ server/start.go
@@ -18,6 +18,7 @@ const (
 	flagWithTendermint = "with-tendermint"
 	flagAddress        = "address"
 	flagTraceStore     = "trace-store"
+	flagPruning        = "pruning"
 )
 
 // StartCmd runs the service passed in, either stand-alone or in-process with
@@ -43,6 +44,7 @@ func StartCmd(ctx *Context, appCreator AppCreator) *cobra.Command {
 	cmd.Flags().Bool(flagWithTendermint, true, "Run abci app embedded in-process with tendermint")
 	cmd.Flags().String(flagAddress, "tcp://0.0.0.0:26658", "Listen address")
 	cmd.Flags().String(flagTraceStore, "", "Enable KVStore tracing to an output file")
+	cmd.Flags().String(flagPruning, "syncable", "Pruning strategy: syncable, nothing, everything")
 
 	// add support for all Tendermint-specific command line options
 	tcmd.AddNodeFlags(cmd)