+++ store/iavlstore_test.go
@@ -266,13 +266,11 @@ func nextVersion(iavl *iavlStore) {
 	iavl.Set(key, value)
 	iavl.Commit()
 }
+
 func TestIAVLDefaultPruning(t *testing.T) {
 	//Expected stored / deleted version numbers for:
 	//numRecent = 5, storeEvery = 3
-	var states = []struct {
-		stored  []int64
-		deleted []int64
-	}{
+	var states = []pruneState{
 		{[]int64{}, []int64{}},
 		{[]int64{1}, []int64{}},
 		{[]int64{1, 2}, []int64{}},
@@ -290,6 +288,39 @@ func TestIAVLDefaultPruning(t *testing.T) {
 		{[]int64{3, 6, 9, 10, 11, 12, 13, 14}, []int64{1, 2, 4, 5, 7, 8}},
 		{[]int64{3, 6, 9, 10, 11, 12, 13, 14, 15}, []int64{1, 2, 4, 5, 7, 8}},
 	}
+	testPruning(t, int64(5), int64(3), states)
+}
+
+func TestIAVLAlternativePruning(t *testing.T) {
+	//Expected stored / deleted version numbers for:
+	//numRecent = 3, storeEvery = 5
+	var states = []pruneState{
+		{[]int64{}, []int64{}},
+		{[]int64{1}, []int64{}},
+		{[]int64{1, 2}, []int64{}},
+		{[]int64{1, 2, 3}, []int64{}},
+		{[]int64{1, 2, 3, 4}, []int64{}},
+		{[]int64{2, 3, 4, 5}, []int64{1}},
+		{[]int64{3, 4, 5, 6}, []int64{1, 2}},
+		{[]int64{4, 5, 6, 7}, []int64{1, 2, 3}},
+		{[]int64{5, 6, 7, 8}, []int64{1, 2, 3, 4}},
+		{[]int64{5, 6, 7, 8, 9}, []int64{1, 2, 3, 4}},
+		{[]int64{5, 7, 8, 9, 10}, []int64{1, 2, 3, 4, 6}},
+		{[]int64{5, 8, 9, 10, 11}, []int64{1, 2, 3, 4, 6, 7}},
+		{[]int64{5, 9, 10, 11, 12}, []int64{1, 2, 3, 4, 6, 7, 8}},
+		{[]int64{5, 10, 11, 12, 13}, []int64{1, 2, 3, 4, 6, 7, 8, 9}},
+		{[]int64{5, 10, 11, 12, 13, 14}, []int64{1, 2, 3, 4, 6, 7, 8, 9}},
+		{[]int64{5, 10, 12, 13, 14, 15}, []int64{1, 2, 3, 4, 6, 7, 8, 9, 11}},
+	}
+	testPruning(t, int64(3), int64(5), states)
+}
+
+type pruneState struct {
+	stored  []int64
+	deleted []int64
+}
+
+func testPruning(t *testing.T, numRecent int64, storeEvery int64, states []pruneState) {
 	db := dbm.NewMemDB()
 	tree := iavl.NewVersionedTree(db, cacheSize)
 	iavlStore := newIAVLStore(tree, numRecent, storeEvery)
@@ -307,6 +338,7 @@ func TestIAVLDefaultPruning(t *testing.T) {
 		nextVersion(iavlStore)
 	}
 }
+
 func TestIAVLNoPrune(t *testing.T) {
 	db := dbm.NewMemDB()
 	tree := iavl.NewVersionedTree(db, cacheSize)
@@ -321,6 +353,7 @@ func TestIAVLNoPrune(t *testing.T) {
 		nextVersion(iavlStore)
 	}
 }
+
 func TestIAVLPruneEverything(t *testing.T) {
 	db := dbm.NewMemDB()
 	tree := iavl.NewVersionedTree(db, cacheSize)