+++ CHANGELOG.md
@@ -12,6 +12,7 @@ BREAKING CHANGES
 FEATURES
 * [baseapp] NewBaseApp now takes option functions as parameters
 * [store] Added support for tracing multi-store operations via `--trace-store`
+* [store] Pruning strategy configurable with pruning flag on gaiad start
 
 BUG FIXES
 * \#1630 - redelegation nolonger removes tokens from the delegator liquid account